refactor: share bind directive processing
